requirements.txt
boto3==1.34.0
python-dotenv==1.0.1
semantic-version==2.10.0